Revert rope back to ttnn and introduce op constraints check for fusing (#6018)
closes #6004

This PR reverts fusing of rope to TTNN. This is simply because there are
a lot of constraints in metal regarding rope fusing which simply cannot
be checked in fusing itself. Instead we propose to use op model get
constraints api to check if rope can be fused. This has some side
effects:

* Rope can only be fused with optimizer on
* Rope tests now require op model which cases golden tests to fail
because of https://github.com/tenstorrent/tt-mlir/issues/5283
https://github.com/tenstorrent/tt-mlir/issues/5909. Problem is that
golden test are opening the device by default and caching it for all the
tests that are being executed, but get constraints api also needs to
open the device and asserts because device is already opened.