[TTNNOpModel] Fix SDPA decode curPosTensor handling (#6123)
### Problem description
In the following
[PR](https://github.com/tenstorrent/tt-mlir/commit/cccd1a388f911436a3f3c85ffac2004a3e5f9e51)￼,
the signature of the SDPA decode op was changed so that curPosTensor
became optional, but this behavior was not updated in TTNNOpModel, which
caused a regression in the LLaMA benchmark.

### What's changed
Updated TTNNOpModel so it correctly handles an optional curPosTensor.

Note: I will merge this as soon as I get approval to fix the regression.

### Checklist
- [ ] New/Existing tests provide coverage for changes