Canonicalize `ttir.update_cache` to `ttir.paged_update_cache` (#6029)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/5745
https://github.com/tenstorrent/tt-mlir/issues/5859

### Problem description
`ttnn.update_cache` requires that all users have the same cache
position. This forces implementations in torch to use left padding. In
addition to that the `update_idx` is a `uint32_t` arg on the ttnn op,
not a tensor. Since we model the op using an integer tensor to carry
this value, we must workaround this by **pulling the tensor off device
and reading the integer from the buffer**. This is not ideal.
### What's changed
`ttnn.paged_update_cache` is also capable of updating a static cache.
This op allows us to provide a different cache position (update_idx) for
each user. This op also accepts an integer tensor as an argument for
this, unlike `ttnn.update_cache`

Since `ttnn.paged_update_cache` has broader functionality, this PR adds
a canonicalization to `ttir.update_cache` to transform it into a
`ttnn.paged_update_cache`, utilizing its static-cache ability.

`ttnn.paged_update_cache` requires its fill value to be in the form `[1,
num_users, num_heads, head_size]` unlike `ttnn.update_cache` which
requires `[1, num_heads, num_users, head_size]` so we must insert a
permute on this operand. It seems as though a permute is already in
place on this operand for most models we compile from `transformers`, so
this extra permute just gets folded away and we end up with one permute
regardless.


If each user does indeed have the same cache position (i.e. the inputs
are left-padded). A single cache position integer tensor will not
suffice. `ttnn.paged_update_cache` _does_ require a cache position
tensor of shape `[batch_size]` so, this tensor must be full of identical
integers.

For some models in `transformers` left padding is simply required, and
issues will arise if you attempt to provide a cache position which is
not a singleton tensor. For this reason, as part of the canonicalization
we will conditionally insert a repeat op on the cache position if it is
a single integer.

### Checklist
- [x] New/Existing tests provide coverage for changes