[D2M] Promote constant tensors to scalars (#5970)
### Ticket
Part 2 of adding scalar support for binary ops. Depends on #5915

### Problem description
Need to convert tensors of a constant to a scalar when possible

### What's changed
Added a pass after `TTIRToD2M` that goes thru all the `d2m.full` ops and
traces to see if they can be turned into constants in the compute ops.
If so, removes the full op and all corresponding values such as
tensor/tile from the generic op
Div scalar has an issue in tt-metal/tt-llk where it does multiplication
instead of division. Issue filed
https://github.com/tenstorrent/tt-mlir/issues/5924 for follow on

### Checklist
- [ ] New/Existing tests provide coverage for changes