[D2M] Handle all shape + stride changes as affine map => view + DMA (#5540)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/3037

### Problem description
We lacked support for lowering what we called a "layout change" in
LowerToLayout--any fundamental change to the alignments, collapse dims,
etc. could not be lowered.

### What's changed
Consolidate grid changes + stride changes into a new class of change
called "mappingChange"-- we inverse the logical -> device mapping on the
input, compose this with the logical -> device mapping on the output,
and therefore map the input to the output as a purely affine-map
transformation. This allows us to lower a broad range of changes through
a single simple mechanism.
We also immediately materialize these views into new tensors via
inserting a genericOp with a DMAOp--this was necessary to avoid race
conditions when evaluating the views (using StreamLayoutOps instead did
not work, due to lack of a folding pattern today).


### Checklist
- [X] New/Existing tests provide coverage for changes