[D2M] Add simple verifier to MetalLayoutAttr (#6174)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/5449

### Problem description
When manually creating MetalLayoutAttrs, it's easy to violate
assumptions which are central to its usage.

### What's changed
Add basic verification s.t. malformed MetalLayoutAttrs get rejected.
Worth noting that we cannot verify anything related to the actual tensor
shape (so if a MetalLayoutAttr is attached a tensor with a mismatched
shape, we can't catch that here), but we can check that the fields of
the attr itself are all internally consistent. Add a negative lit test
to ensure coverage too.

### Checklist
- [X] New/Existing tests provide coverage for changes