[Optimizer] Temporary workaround for divide-by-zero in matmul kernel (#6169)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-forge/issues/710)

### Problem description
When running the segformer model on the latest tt-xla with optimizer
enabled, compilation crashes in the `TTNNOptimizer` pass while computing
constraints for a `MatmulOp`.

The root cause is that the matmul kernel currently uses the logical
shape instead of the padded shape to calculate the number of tiles for
block-sharded input. When the last dimension is less than 32, this
results in 0 tiles, leading to a divide-by-zero floating-point
exception.

### What's changed
This PR introduces a temporary workaround to prevent crashes. 

It adds additional validation to `MatmulOp::getOpConstraints` in
`TTNNOpModelInterface`, along with a new test that verifies the fix has
not yet been uplifted in tt-metal.

The test currently expects the matmul operation to crash. When it starts
failing, that will indicate that [this fix on
tt-metal](https://github.com/tenstorrent/tt-metal/commit/c579ac2ad7b1df0ab4ab35f308eb1e67e43751ca)
is uplifted and this workaround should be removed.

### Checklist
- [x] New/Existing tests provide coverage for changes