[cmake] set metal build dir properly (#6076)
### Problem description
When switching from release build to debug build, I got the error that
some variables are unused. It was because the compiler expanded the code
to `(void) 0` when `SPDLOG_ACTIVE_LEVEL` is greater than DEBUG. The log
level was controlled by the cmake variable `TT_METAL_ENABLE_LOGGING`
which was set by the release build, and reused in the subsequent debug
build. This fix uses different build directories for debug and release
builds, so that debug builds won't reuse release builds' log level
settings.

Please refer to the PR for the full description.

### What's changed
- Introduced the cmake variable `TTMETAL_BUILD_DIR`.
- Set different value of `TTMETAL_BUILD_DIR` for release and debug
builds.
- Pass `TTMETAL_BUILD_DIR` as an environment variable to ttrt builds.

### Checklist
- [X] Builds passed when switching from release to debug and vice versa.