Force host memory layout for const-eval inputs (#6198)
### Ticket
#6048 

## Problem description
Const-eval inputs are placed in DRAM by default. Since const-eval
results are also stored in DRAM, this effectively means that if the
deallocation of original inputs doesn't happen (because inputs are
retained on the frontend), memory footprint is twice the actually
required size.

OOMs which can be pin-pointed to this exact issue have been observed on
some models.

## What's changed
- Force const-eval function arguments to system memory
- Force function arguments which get consumed by `load_cached` op to
system memory
- Extend the logic for handling `ttir.mesh_shard` input and output
layouts

### Checklist
- [x] Revert temporary nanobind dependencies flag
- [x] New/Existing tests provide coverage for changes