Fixing errors with pad for ttnn.avgPool2d (#6245)
### Ticket
closes https://github.com/tenstorrent/tt-mlir/issues/5785

### Problem description
Pooling op gets fused with pad op, but tt-metal has constrain that pad
must be smaller than half of the kernel size. Verifier only checks that
input dimensions are greater than kernel dimensions, disregarding
padding.

### What's changed
If padding is greater than allowed, added decomposition to padOp +
poolingOp. It is not done as a special case when it should not be fused
because it is mathematically correct, but this is a hardware constraint.
Fixed verifier logic.

### Checklist
- [x] New/Existing tests provide coverage for changes