Fix for BH coord translation issue
Translated coordinates in BH are not contiguous so we cannot use a pair of translation offsets to convert
coordinates from LOGICAL to TRANSLATED system. We instead use the worker_logical_col_to_virtual_col and
worker_logical_row_to_virtual_row tables (which are global arrays on device) to do this conversion.