[CMAKE] Alias Python_VERSION to Python3_VERSION for nanobind >= 2.10 (#6266)
nanobind introduced `Python_VERSION` to their cmake config when removing
Python 3.8 support:
https://github.com/wjakob/nanobind/commit/b50c3d80bab4f16ee57d680cab113925a7284880#diff-fed8537f94da6b746202926961e2c62ea9386778d758d293e7a24e2011577c64R7

As an aside, I think I might be able to re-work the Triton build process
to remove the need to add these aliases in the future. But that needs a
little more time. For now, this lets Triton update nanobind to 2.10.2
when pulling in `tt-mlir` for the Tenstorrent backend.