Fix for BH coord translation issue (#6305)
### Ticket
Link to Github Issue:
https://github.com/tenstorrent/tt-mlir/issues/6248,
https://github.com/tenstorrent/tt-mlir/issues/5341

### Problem description
Translated coordinates in BH are not contiguous so we cannot use a pair
of translation offsets to convert coordinates from LOGICAL to TRANSLATED
system.

### What's changed
We instead use the worker_logical_col_to_virtual_col and
worker_logical_row_to_virtual_row tables (which are global arrays on
device) to do this conversion.

### Checklist
- [x] New/Existing tests provide coverage for changes