[D2M] Add bfp8 typecast tests (#5985)
Added builder tests for bfp8 typecasting through the
GlobalDataFormatConversion pass. Bfp8 is not currently supported in
builder and typecast-to-bfp8 operations cannot be explicitly created.

Excluded ops with unit attribute `preserveDataFormat` from the data
conversion pass. This was to be able to tests a chain ops with (mostly
bfp8) typecasts. The resulting IR is not the cleanest, but it works for
the purpose of testing the robustness of multiple typecasts intertwined
with other ops.

Added `D2M_SkipOpEltwiseFusionTrait ` to typecasts. The current fusion
passes break when typecasts are present.