[D2M] Cleanup tensor_layout parsing in MetalLayoutAttr (#6370)
### Ticket
NA

### Problem description
The `memory_layout` field in MetalLayoutAttr is annoying me because it
is default valued, which implies it is not printed when it is `sharded`.
However, this means we cannot trivially re-ingest IR, I get an error
when trying to parse IR dumped with this attr left blank like `l1,,
index_map = ...` instead of `l1, sharded, index_map = ...`. And it's
ugly.

Error example:
```
tmp.mlir:1:147: error: expected keyword for TTCore Tensor Memory Layout
#layout = #ttcore.metal_layout<logical_shape = 32x32, dim_alignments = 64x64, collapsed_intervals = dense<[[0, 1]]> : tensor<1x2xi64>, neginf, l1,, index_map = map(4)>
                                                                                                                                                  ^
tmp.mlir:1:147: error: failed to parse TTCore_MetalLayoutAttr parameter 'memory_layout' which is to be a `TensorMemoryLayout`
#layout = #ttcore.metal_layout<logical_shape = 32x32, dim_alignments = 64x64, collapsed_intervals = dense<[[0, 1]]> : tensor<1x2xi64>, neginf, l1,, index_map = map(4)>
                                                                                                                                                  ^
```

### What's changed
Use AttrParameter for TensorMemoryLayout instead--this will print
properly, re-parse properly, etc. Updated effected tests, add new test
to confirm we can roundtrip properly.

### Checklist
- [X] New/Existing tests provide coverage for changes