added d2m support for reshape and tests (#6378)
### Problem description
Need to support reshape in d2m 

### What's changed
Added support for changing the rank of tensors in
`projectLogicalMapToUnitDeviceSpace` which is needed for reshape and
`reshapeLogicalMap` which is the affine map for reshape transformations.
Assumes minimum rank is 2 for now.
Added builder tests for tile aligned tensors, random shaped tensors and
tensors found in llama 3.2 3b. Lots of these are skipped for various
reasons, such as compile time #6375, 1d support #6376, and noc unaligned
reads/writes #6377.

Had an initial hacky solution for 1d tensors by adding a leading 1 to
the metal layout attribute so it would pass the assertion but didn't get
to handling all the edge cases so leaving that aside and marking the 1d
llama tests as skip for now. I could pick this up when I'm back or we
could try and support this natively in the pipeline.

### Checklist
- [ ] New/Existing tests provide coverage for changes