D2M: fix missed alignments for certain TMs (#6411)
Closes #6268

### Bug 1
When the TTIRToD2M pass creates StreamLayoutOps for TMs, it uses unit
grids with default alignments and no collapsing.

In the GridSelection pass, a grid is applied to the storage of the
StreamLayoutOps and it's possible to get an higher alignment requirement
than the default.

Example: apply rearrange with collapse `z y x -> y (z x)` on a `33x2x8`
tensor to get a `2x264` tensor on a `8x8` grid will endup with a
`32x256` alignment requirement on the stream's storage.

When the GridSelection pass creates a new stream storage on an optimal
grid, a reblock map from the new optimal storage to the old unit grid
shape is prepended to the stream's affine map, so the semantics of the
stream is preserved.

However, this reblock map makes no sense when these two shapes have
different dimAlignments: their volumes are different and their strides
are incompatiable and so the index transformation is incorrect.

Solution: when an alignment change is detected, we align-up the old unit
grid shape with the new alignments, and calculate the correct reblock
map using it.

Since alignment paddings are applied at the end of the dimensions, this
align-up "aligns" the first step and doesn't affect the second step of
the index transformation chain
`optimal_output->unit_output->tm_map->input_index_map`.

### Bug 2
When applying rearrangement `z y x -> (z y) x` to `4x85x1055` to get
`340x1055`, the stream's unit grid storage's shape would be `352x1056`
and dimAlignments would be `256x256` from the start, so we won't
encounter the bug above.

The missed alignment problem manifests later, in the LowerToLayout pass,
and on the stream's input this time.

Since the GridSelection pass selected optimal (virtual) grid `2x3x8` for
the input tensor, its device shape would be `2x3x8x2x32x160` with
alignment `1x32x256` and no collapse.

When we lower the ToLayoutOp for the host to device transfer of this
input we trigger the virtual bounce and collapse the optimal grid to be
`6x8`.

This 3D to 2D collapse leads to a bigger alignment of `256x32x256` and
changes the shape of the bounce tensor, and in this particular case,
made it indivisible by grid `6x8`.

Solution: when an alignment change is detected, bounce the H2D transfer
through an interleaved DRAM tensor on the unit grid, and let the
DRAM->L1 lowering step take over.

### Misc
- Add an assert in reblock map calculation to catch obvious nonsensical
inputs.
- Correct confusing variable names.

---------

Signed-off-by: wenbinlyuTT <wenbinlyu@tenstorrent.com>