Use top-down traversal during Sdy CCLs to SHLO CCLs pattern rewriter (#6421)
### Ticket
Fixes https://github.com/tenstorrent/tt-mlir/issues/6157

### Problem description
The current pattern rewriter in `convertShardyCCLToStableHLOCCL()`
within the `UpdateGlobalToLocalShapesPass` uses bottom-up traversal.
This leads to cases where, when the result of a Sdy CCL op is fed as
input to another Sdy CCL op, the rewriter fails to update the shape
correctly and causes the pass to fail. See attached ticket for more
context.

### What's changed
- Use top-down traversal instead so that the shapes are updated in the
correct order.

### Checklist
- [x] New/Existing tests provide coverage for changes