[consteval] make const-eval cache global (#6434)
Currently, consteval cache (i.e. `TensorCache`) is created per
flatbuffer binary. It turns out that there are cases (LLMs) where we
have multiple identical const-eval graphs present in multiple binaries
(modules), e.g. prefill & decode. For models that barely fit on a single
chip (llama-8b, qwen-8b, etc.) this causes us to hit OOM; since the
cached results for the identical const-eval graphs are stored two times,
one in the prefill's cache and the other in the decode's cache.

This change modifies the const-eval scheme such that it is possible to
have one unique (global) cache - reducing the memory footprint.

To enable this, we need a way to identify identical const-eval graphs in
the runtime. After investigating some options, I've chosen to go with
hashing the IR with `sha256` and storing that hash value in the
flatbuffer binary for each const-eval function.

NOTE: trying to identify mathematically equivalent IRs is not part of
the scope of this change - here we only care for identical IRs.

The hash is computed on the dumped IR string, which is created in the
following way:
- the function name (e.g. "main_const_eval_123") is temporarily replaced
with "anonymous_const_eval" so that it doesn't affect the computed hash
- the IR is printed using `useLocalScope()` and `printGenericOpForm()`.

To help detect false positives (semantically different functions having
the same hash), we could implement a runtime debug check which
recomputes output tensors each time and verifies that the same tensors
are in the cache - however, I would prefer to do it as a follow up if
needed. Having false positives seems unlikely, when operating under the
assumption that all the info needed for execution is contained within
ops and their attributes, which in turn all get dumped as IR text.

Once we have the hash, then we can use it as part of the key in the
cache map in runtime. The cache key contains of:
- id of the device on which we executed the const-eval function
- function hash
- vector of tensor versions - which uniquely identify which input
tensors were used to compute the cached result.

In runtime, the `TensorCache` class is now `GlobalTensorCache` which is
a singleton. It is moved into `ttnn` namespace, since it is designed and
used only in `ttnn` backend.

A callback mechanism ties cache entry lifetimes to the lifetimes of its
input tensors, removing the need for explicit cache management in most
cases. We were having issues with the previous design since the cache
was tied to the binary; so if you wanted to cache binaries for reuse,
you would need to manually clean up its cache.

Unit tests and end-to-end python test are added.

Closes #3888