Mismatch in dimensions for embeddingbw (#6464)
### Ticket
closes https://github.com/tenstorrent/tt-mlir/issues/6446

### Problem description
Conversion from stableHLO.scatterOp to ttnn.embeddingBW op requires
padding if shape is not aligned with tile width. However, scatter can be
mapped even if dimensions are 2D. Because stableHLO represents indices
with vector of size 1, embeddingBW incorrectly deciphers what is the
dimension that should be padded and results in a crash.

### What's changed
Added a reshape to conversion if it is necessary. 
### Checklist
- [x] New/Existing tests provide coverage for changes